requirements.txt
Flask==2.3.3
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
werkzeug==2.3.7
ollama==0.1.7